services/risingwave/pyproject.toml
[project]
name = "risingwave"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.38",
    "sqlalchemy-risingwave>=1.1.1",
]
